Syntax Reference

SymbolServer.get_cache_path

SymbolServer.get_pkg_path

SymbolServer.isinmanifest

SymbolServer.isinproject

SymbolServer.load_package_from_cache_into_store!

SymbolServer.manifest

SymbolServer.maybe_fix_path

SymbolServer.maybe_fixup_stdlib_path

SymbolServer.maybe_getfield

SymbolServer.project

SymbolServer.read_manifest

SymbolServer.remove_non_general_pkgs!

SymbolServer.split_module_names

SymbolServer.validate_disc_store

SymbolServer.get_cache_path — Method

get_cache_path(manifest, uuid)

Returns a vector containing the cache storage path for a package structured: [folder, folder, file].

SymbolServer.get_pkg_path — Method

get_pkg_path(pkg::Base.PkgId, env, depot_path)

Find out where a package is installed without having to load it.

SymbolServer.isinmanifest — Function

isinmanifest(context, package::Union{String,UUID})

Checks whether a package is in the manifest of a given context, e.g. is either directly loadable or is a dependency of an loadable package.

SymbolServer.isinproject — Function

isinproject(context, package::Union{String,UUID})

Checks whether a package is in the dependencies of a given context, e.g. is directly loadable.

SymbolServer.load_package_from_cache_into_store! — Function

load_package_from_cache_into_store!(ssp::SymbolServerInstance, uuid, store)

Tries to load the on-disc stored cache for a package (uuid). Attempts to generate (and save to disc) a new cache if the file does not exist or is unopenable.

SymbolServer.manifest — Method

manifest(c::Pkg.Types.Context)

Retrieves the UUID -> PackageEntry map from the manifest of a Context.

SymbolServer.maybe_fix_path — Method

path = maybe_fix_path(path)

Return a normalized, absolute path for a source file path.

SymbolServer.maybe_fixup_stdlib_path — Method

path = maybe_fixup_stdlib_path(path::String)

Return path corrected for julia issue #26314 if applicable. Otherwise, return the input path unchanged.

Due to the issue mentioned above, location info for methods defined one of Julia's standard libraries are, for non source Julia builds, given as absolute paths on the worker that built the julia executable. This function corrects such a path to instead refer to the local path on the users drive.

SymbolServer.maybe_getfield — Method

maybe_getfield(k::Symbol , m::ModuleStore, server)

Try to get k from m. This includes: unexported variables, and variables exported by modules used within m.

SymbolServer.project — Method

project(c::Pkg.Types.Context)

Retrieves the project of a Context.

SymbolServer.read_manifest — Method

read_manifest(manifest_filename)

Read the manifest from the path and return the UUID -> PackageEntry map. If the file can't be read, return nothing.

SymbolServer.remove_non_general_pkgs! — Method

remove_non_general_pkgs!(pkgs)

Removes packages that aren't going to be on the symbol cache server because they aren't in the General registry. This avoids leaking private package name & uuid pairs via the url requests to the symbol server.

If the General registry cannot be found packages cannot be checked, so all packages will be removed.

SymbolServer.split_module_names — Method

split_module_names(m::Module, allns)

Return two lists of names accessible from calling getfield(m, somename). The first contains those symbols returned by Base.names(m, all = true). The second contains all others, including imported symbols and those introduced by the using of modules.

SymbolServer.validate_disc_store — Method

validate_disc_store(store_path, manifest)

This returns a list of non-jll packages in the manifest that don't have caches on disc.