Syntax Reference

SymbolServer.isinmanifest

SymbolServer.isinproject

SymbolServer.load_package_from_cache_into_store!

SymbolServer.manifest

SymbolServer.maybe_fix_path

SymbolServer.maybe_fixup_stdlib_path

SymbolServer.maybe_getfield

SymbolServer.project

SymbolServer.split_module_names

SymbolServer.isinmanifest — Function

isinmanifest(context, package::Union{String,UUID})

Checks whether a package is in the manifest of a given context, e.g. is either directly loadable or is a dependency of an loadable package.

SymbolServer.isinproject — Function

isinproject(context, package::Union{String,UUID})

Checks whether a package is in the dependencies of a given context, e.g. is directly loadable.

SymbolServer.load_package_from_cache_into_store! — Method

load_package_from_cache_into_store!(ssp::SymbolServerInstance, uuid, store)

Tries to load the on-disc stored cache for a package (uuid). Attempts to generate (and save to disc) a new cache if the file does not exist or is unopenable.

SymbolServer.manifest — Method

manifest(c::Pkg.Types.Context)

Retrieves the manifest of a Context.

SymbolServer.maybe_fix_path — Method

path = maybe_fix_path(path)

Return a normalized, absolute path for a source file path.

SymbolServer.maybe_fixup_stdlib_path — Method

path = maybe_fixup_stdlib_path(path::String)

Return path corrected for julia issue #26314 if applicable. Otherwise, return the input path unchanged.

Due to the issue mentioned above, location info for methods defined one of Julia's standard libraries are, for non source Julia builds, given as absolute paths on the worker that built the julia executable. This function corrects such a path to instead refer to the local path on the users drive.

SymbolServer.maybe_getfield — Method

maybe_getfield(k::Symbol , m::ModuleStore, server)

Try to get k from m. This includes: unexported variables, and variables exported by modules used within m.

SymbolServer.project — Method

project(c::Pkg.Types.Context)

Retrieves the project of a Context.

SymbolServer.split_module_names — Method

split_module_names(m::Module, allns)

Return two lists of names accessible from calling getfield(m, somename). The first contains those symbols returned byBase.names(m, all = true). The second contains all others, including imported symbols and those introduced by theusing` of modules.